refactor: extract `isListBlock` into its own util
