+++ packages/editor/src/behaviors/behavior.core.block-objects.ts
@@ -10,8 +10,8 @@ import {getLastChild} from '../traversal/get-last-child'
 import {getLeaf} from '../traversal/get-leaf'
 import {getNode} from '../traversal/get-node'
 import {getSibling} from '../traversal/get-sibling'
-import {isListBlock} from '../utils/parse-blocks'
 import {isEmptyTextBlock} from '../utils/util.is-empty-text-block'
+import {isListBlock} from '../utils/util.is-list-block'
 import {raise} from './behavior.types.action'
 import {defineBehavior} from './behavior.types.behavior'
